requirements.txt
discord.py>=2.4.0
aiohttp>=3.8.0
colorama>=0.4.6